# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZenMon API Documentation",
        "description": "Monitoring Application - Alternative to Zabbix/Nagios",
        "contact": {
            "email": "admin@zenmon.local"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "ZenMon API Server"
        }
    ],
    "paths": {
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get list of alerts with filtering (UC34)",
                "description": "Returns paginated list of alerts with filtering options for dashboard and alert history",
                "operationId": "getAlertsList",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by alert status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Active",
                                "Acknowledged",
                                "Closed"
                            ]
                        }
                    },
                    {
                        "name": "host_id",
                        "in": "query",
                        "description": "Filter by host ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "alert_level",
                        "in": "query",
                        "description": "Filter by alert level",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Warning",
                                "Critical"
                            ]
                        }
                    },
                    {
                        "name": "metric_type_id",
                        "in": "query",
                        "description": "Filter by metric type",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Start date for filtering",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "End date for filtering",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of alerts per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Alert"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/alert-thresholds": {
            "get": {
                "tags": [
                    "Alert Thresholds"
                ],
                "summary": "Get list of alert thresholds (UC40)",
                "description": "Returns list of alert thresholds with filtering options",
                "operationId": "getAlertThresholdsList",
                "parameters": [
                    {
                        "name": "host_id",
                        "in": "query",
                        "description": "Filter by host ID (null for global thresholds)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "metric_type_id",
                        "in": "query",
                        "description": "Filter by metric type ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "global_only",
                        "in": "query",
                        "description": "Show only global thresholds (host_id = null)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlertThreshold"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alert Thresholds"
                ],
                "summary": "Create new alert threshold (UC40)",
                "description": "Create new threshold configuration for alerts. Set host_id to null for global thresholds.",
                "operationId": "createAlertThreshold",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "metric_type_id",
                                    "warning_threshold",
                                    "critical_threshold"
                                ],
                                "properties": {
                                    "host_id": {
                                        "description": "Host ID (null for global threshold)",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "metric_type_id": {
                                        "description": "Metric Type ID",
                                        "type": "integer"
                                    },
                                    "warning_threshold": {
                                        "description": "Warning threshold value",
                                        "type": "number"
                                    },
                                    "critical_threshold": {
                                        "description": "Critical threshold value",
                                        "type": "number"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Threshold created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AlertThreshold"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Threshold already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/connection-statuses": {
            "get": {
                "tags": [
                    "Connection Status"
                ],
                "summary": "Get list of connection statuses (UC23)",
                "description": "Returns paginated list of connection statuses with filtering",
                "operationId": "getConnectionStatusesList",
                "parameters": [
                    {
                        "name": "host_id",
                        "in": "query",
                        "description": "Filter by host ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by connection status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Online",
                                "Offline",
                                "Unknown"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter from date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter to date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "latest_only",
                        "in": "query",
                        "description": "Get only latest status per host",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ConnectionStatus"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Connection Status"
                ],
                "summary": "Store new connection status (UC23)",
                "description": "Record new connection status check result",
                "operationId": "storeConnectionStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "host_id",
                                    "status",
                                    "last_check_date"
                                ],
                                "properties": {
                                    "host_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Online",
                                            "Offline",
                                            "Unknown"
                                        ],
                                        "example": "Online"
                                    },
                                    "last_check_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-06-22T10:30:00Z"
                                    },
                                    "error_message": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "response_time": {
                                        "type": "integer",
                                        "example": 145,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Connection status recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ConnectionStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Alert": {
                "title": "Alert",
                "description": "Alert model",
                "properties": {
                    "alert_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "host_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "metric_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "alert_level": {
                        "type": "string",
                        "enum": [
                            "Warning",
                            "Critical"
                        ],
                        "example": "Critical"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Active",
                            "Acknowledged",
                            "Resolved"
                        ],
                        "example": "Active"
                    },
                    "alert_message": {
                        "type": "string",
                        "example": "CPU usage exceeded critical threshold"
                    },
                    "current_value": {
                        "type": "number",
                        "format": "float",
                        "example": 95.5
                    },
                    "threshold_value": {
                        "type": "number",
                        "format": "float",
                        "example": 90
                    },
                    "created_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-22T10:30:00Z"
                    },
                    "acknowledged_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "acknowledged_by": {
                        "type": "integer",
                        "nullable": true
                    },
                    "resolved_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "host": {
                        "properties": {
                            "host_id": {
                                "type": "integer"
                            },
                            "host_name": {
                                "type": "string"
                            },
                            "ip_address": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "metric_type": {
                        "properties": {
                            "metric_type_id": {
                                "type": "integer"
                            },
                            "metric_name": {
                                "type": "string"
                            },
                            "unit": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}